fix: seaprate storage functions
