+++ functions/src/handler/https.ts
@@ -1,9 +1,13 @@
-import express from 'express'
+import { FunctionInfo, Logger } from '../globals'
+import { fetchDqx9mbrpz1jhx } from '../utils/fetch'
+import { getFilename } from '../utils/storage'
 import cors from 'cors'
+import express from 'express'
+import { initializeApp } from 'firebase-admin/app'
+import { getStorage } from 'firebase-admin/storage'
 import * as functions from 'firebase-functions/v1'
 
-import { fetchDqx9mbrpz1jhx } from '../utils/fetch'
-import { Logger } from '../globals'
+initializeApp()
 
 const corsHandler = cors({ origin: true })
 
@@ -24,7 +28,19 @@ export const dqx9mbrpz1jhxHandler = (
       for (const elem of urls) {
         // Iterate files in urls.
         const url = new URL(elem)
-        await fetchDqx9mbrpz1jhx(url)
+        const bucket = getStorage().bucket(FunctionInfo.BUCKET_NAME)
+        const filename = getFilename(url)
+        if (!filename) {
+          Logger.warn('A file name must be specified. : ', filename)
+          return
+        }
+        const file = bucket.file(filename)
+        const [exists] = await file.exists()
+        if (exists) {
+          return
+        }
+
+        await fetchDqx9mbrpz1jhx(url, file)
       }
       res.sendStatus(201)
     } else {
